fixed problems with posts being moved into a new directory
