+++ Tabloid/client/src/components/posts/PostList.js
@@ -1,6 +1,6 @@
 import React, { useContext, useEffect } from "react";
-import { PostContext } from "../providers/PostProvider";
-import Post from "./posts/Post";
+import { PostContext } from '../../providers/PostProvider';
+import Post from "./Post";
 
 const PostList = () => {
   const { posts, getAllPosts } = useContext(PostContext);
