+++ Tabloid/client/src/components/posts/UserPosts.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useContext, useState } from "react";
 import { useParams } from "react-router-dom";
-import { PostContext } from '../providers/PostProvider';
-import Post from "./posts/Post";
+import { PostContext } from '../../providers/PostProvider';
+import Post from "./Post";
 
 const UserPosts = () => {
